api/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "willow-api"
version = "0.1.0"
authors = [
  {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "package_description"
dependencies = [
  "pydantic", "openai", "fastapi", "uvicorn[standard]", "streamed-yaml", "python-dotenv",
  "aiosqlite",
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}

[project.urls]
repo = "https://github.com/moveread/REPO.git"

# [project.optional-dependencies]
# test = [
#   "pytest < 5.0.0",
#   "pytest-cov[all]"
# ]